requirements.txt
requests==2.27.1
pytest==6.2.5
allure-pytest
PyYAML==6.0
pytest-html
